[Analysis] Move SimplifyQuery into separate header (NFC)
To allow reusing it between InstructionSimplify and ValueTracking.
